requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
fastapi==0.110.2
greenlet==3.0.3
h11==0.14.0
idna==3.7
packaging==24.0
pipdeptree==2.18.1
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyMySQL==1.1.0
setuptools==68.2.2
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
wheel==0.41.2
jose==1.0.0
python-multipart==0.0.5
PyJWT==2.6.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
passlib[bcrypt]==1.7.4